Bump dependency numbers to resolve install issues
This fixes #20 which was caused by `react-scripts` installation. In
bumping the version number, another error appeared. I'm not sure if it
was moving to a new version of Node on my machine or if it's an error
from one of the new dep versions. In any case, the changes to the .js
files and the new symlink from src/data.json -> data/data.json is
because imported sources have to be in the src directory now.